Update inlet code
Somewhat large update to the inlet code to change the sensor yaml file
format as well as remove subscription code. This changes the inlet to be
purely a publisher.

The new yaml format allows for root specification and now is formatted
to better encapsulate the "location" of the sensor as part of the topic.