Add kube.yaml file
Podman is designed to run with Dockerfile and kubernetes yaml files for
deployment. The previous docker-compose file is kinda janky as podman is
not a drop-in replacement for docker (and podman-compose is a 3rd party
tool that's kinda janky as well).

kube.yaml uses a kubernetes format for declaring the pod. This can be
run by using `podman play kube kube.yaml`.