Add common crate
A common crate to hold share object definitions now exists in the
workspace. This crate should only hold definitons of structs and enums
that most likely would be shared between each of the crates. Not all
scripts have been modified to use this new common crate. This commit
only modifies part of ingest to use common.