Update create_pod.sh to add db-api and logs
db-api is now required to run the entire platform. This adds db-api to
the create_pod.sh script, meaning it now exists in the pod.

Logging was also added to ingest and db-api, so this adds support for
the volume mounts and the env variables. Also adds an if-statement to
create_pod.sh that creates the pod_logs directory if it doesn't already
exist.