Move router construction to app() in endpoints.rs
The router does not need to be defined in the main function. It is
better off being made in a function within endpoint.rs. This prevents
the endpoints and the logic around them from being exposed to the main
function.

This is also helping pave the way for tests. We need to recreate the app
in every test, so having a reusuable function is super helpful.