Add testing to db-api
Large addition that includes testing for all endpoints in db-api. This
uses SQLx's testing facilities, including auto migration and fixtures.

For running the tests, see the db-api readme for instructions. The tests
require that a Postgres database be running. The tests will
automatically handle adding the required tables and data to the database.