Refactor ingest and add tests
Ingest previously was previously contained entirely in a single binary.
From a code management perspective, this isn't a great idea. I took a
similar approach to how db-api was broken into multiple files and
decided to make an new actors.rs file. Each of the main functions
essentially works as an actor, with communication via channels.

This also adds the beginning of testing for ingest. This is one of the
more difficult parts of the chain to test due to its connection with the
surrounding services. The ingest aspect requires a broker for the system
to connect to, while the calculation/posting aspect requires the db-api
to post new data to.

Still working out how exactly to test this module. I'd like to avoid
mocking every HTTP connection with a fake mock as those can get stale
very quickly. Full integration testing may be the fastest way to get
testing up and running. The only test actually added with this commmit
is the handle_data() test. This function does not need any external
connection, so its easy to test.