Update common to fix issue with test and add comments
Common hadn't been touched since the refactor of ingest. This cleans up
some the unneeded code as well as finishes the tests that never were
completed. We are now tested to ensure that Reading and Sensors properly
serialize into the correct JSON strings.

This also adds some comments on the objects. Nothing too big.