Cleanup inlet and add readme
inlet was almost untouched since the beginning of this codebase. This
commit goes through and does a second pass at small improvements to the
publishing loop as well as better uses the objects that already exist.

A readme explaining the tool and how it works is also now available.
This includes a description of the config file.